+++ R/reexport.R
@@ -24,6 +24,6 @@ khroma::palette_shape
 #' @export
 khroma::palette_line
 
-#' @importFrom khroma palette_size_range
+#' @importFrom khroma palette_size_sequential
 #' @export
-khroma::palette_size_range
+khroma::palette_size_sequential